+++ src/app/commands/poap/POAP.ts
@@ -18,12 +18,19 @@ import { GuildMember } from 'discord.js';
 import ModifyPOAP from '../../service/poap/config/ModifyPOAP';
 import StatusPOAP from '../../service/poap/config/StatusPOAP';
 import { command } from '../../utils/SentryUtils';
+import allowedServers from '../../service/constants/allowedServers';
 
 export default class POAP extends SlashCommand {
 	constructor(creator: SlashCreator) {
 		super(creator, {
 			name: 'poap',
 			description: 'Receive a list of all attendees in the specified voice channel and optionally send out POAP links.',
+			throttling: {
+				usages: 10,
+				duration: 1,
+			},
+			defaultPermission: true,
+			guildIDs: allowedServers,
 			options: [
 				{
 					name: 'config',
@@ -205,11 +212,6 @@ export default class POAP extends SlashCommand {
 					],
 				},
 			],
-			throttling: {
-				usages: 10,
-				duration: 1,
-			},
-			defaultPermission: true,
 		});
 	}